requirements.txt
# requirements.txt  ─ after the fix
git+https://github.com/langchain-ai/langchain-mcp-adapters.git#egg=langchain_mcp_adapters
langchain-openai>=0.1.8
langgraph>=0.0.51
fastmcp>=2.2.5
langchain>=0.1.20
openai>=1.23.0
oci==2.141.1
langchain_core==0.3.29
langchain-community==0.3.14
nemoguardrails
langsmith
streamlit
redis
